tests/submission/W3C/domhtml/HTMLTableElement24.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - TABLE</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-table-element.html#dom-table-deletetfoot" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<table align="center" summary="Table 1">
<tbody><tr>
<th>Id</th>
<th>Name</th>
<th>Position</th>
<th>Salary</th>
</tr>
</tbody></table>
<table align="center" bgcolor="#ff0000" border="4" cellpadding="2" cellspacing="2" frame="border" id="table-1" rules="all" summary="HTML Control Table" width="680">
<caption align="top">Table Caption</caption>
<thead align="center" valign="middle">
<tr align="center" bgcolor="#00FFFF" valign="middle">
<th id="header-1">Employee Id</th>
<th abbr="maiden" align="center" axis="center" bgcolor="#00FFFF" colspan="1" headers="header-1" height="50" id="header-2" nowrap="nowrap" rowspan="1" scope="col" valign="middle" width="100">Employee Name</th>
<th>Position</th>
<th>Salary</th>
<th>Gender</th>
<th>Address</th>
</tr>
</thead>
<tfoot align="center" valign="middle">
<tr>
<th>next page ...</th>
<th>next page ...</th>
<th>next page ...</th>
<th>next page ...</th>
<th>next page ...</th>
<th>next page ...</th>
</tr>
</tfoot>
<tbody align="center" valign="middle">
<tr>
<td abbr="maiden2" align="center" axis="center" bgcolor="#FF0000" colspan="1" headers="header-2" height="50" id="Table-3" nowrap="nowrap" rowspan="1" scope="col" valign="middle" width="175">EMP0001</td>
<td headers="header-2">Margaret Martin</td>
<td>Accountant</td>
<td>56,000</td>
<td>Female</td>
<td>1230 North Ave. Dallas, Texas 98551</td>
</tr>
<tr>
<td>EMP0002</td>
<td>Martha Raynolds</td>
<td>Secretary</td>
<td>35,000</td>
<td>Female</td>
<td>1900 Dallas Road Dallas, Texas 98554</td>
</tr>
</tbody>
</table>
<table summary="Table 3">
<tbody>
<tr>
<td>
</td>
</tr>
</tbody>
<tbody>
<tr>
<td>
</td>
</tr>
</tbody>
<tbody>
<tr>
<td>
</td>
</tr>
</tbody>
</table>


<script>
var domtest = async_test("The deleteTFoot() method deletes the footer from the table.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var rowsnodeList;
      var testNode;
      var vsection1;
      var vsection2;
      var vrows;
      var doc;
      var result = new Array();

      expectedResult = new Array();
      expectedResult[0] = 4;
      expectedResult[1] = 3;

      doc = window.document;
      nodeList = doc.getElementsByTagName("table");
      assert_equals(nodeList.length,3);
testNode = nodeList.item(1);
      vsection1 = testNode.tFoot;

      assert_not_equals(vsection1, null);
rowsnodeList = testNode.rows;

      vrows = rowsnodeList.length;

      result[result.length] = vrows;
testNode.deleteTFoot();
      vsection2 = testNode.tFoot;

      rowsnodeList = testNode.rows;

      vrows = rowsnodeList.length;

      result[result.length] = vrows;
assert_equals_list(result,expectedResult);
       
    domtest.done();
}));

</script>
<p>The deleteTFoot() method deletes the footer from the table. 

    The deleteTFoot() method will delete the TFOOT Element from the
    second TABLE element.  First make sure that the TFOOT element exists
    and then count the number of rows.  After the TFOOT element is
    deleted there should be one less row.</p>
<div id="log"></div>
</body></html>
